Automating Feedback Form Data Analysis With Python Cash Study
The feedback form automated data analysis case study is for an institute. The institute offers numerous online courses. The institute provides a feedback form to all participants at the end of the course. This feedback form includes questions about faculty skills, course usefulness, and whether or not sufficient course materials are supplied. The institute asked these kinds of multiple-choice questions. All participating members are expected to complete the feedback form. The feedback form is presented after each course, and the questions are the same each time. After gathering replies from all members, the analysis and visualization are done manually in MS Excel, which can take up to four hours. This project is supervised by a technical officer, and the institute wants software that automatically processes data, analyses it, and visualizes it based on feedback forms.